fiasco_api/tags/serializers.py
from rest_framework import serializers

from .models import Tag


class TagSerializer(serializers.ModelSerializer):
    owner = serializers.HiddenField(default=serializers.CurrentUserDefault())

    class Meta:
        model = Tag
        fields = '__all__'
