requirements.txt
ipython
ipdb
django
django-rest-framework 
django_extensions
django_rest_auth
psycopg2
django-cors-headers
django-configurations